% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmiknn.R
\name{cmiknn}
\alias{cmiknn}
\title{Conditional mutial information-k-nearest-neighbors conditional independence test}
\usage{
cmiknn(
  X,
  Y,
  Z,
  knn = 0.15,
  shuffle_neighbors = 5,
  sig_samples = 500,
  transform = "ranks"
)
}
\arguments{
\item{X}{A nxp-matrix of n p-dimensional vectors}

\item{Y}{A nx1-matrix}

\item{Z}{A nxq-matrix of n q-dimensional confounder (optional)}

\item{knn}{Number of nearest-neighbors which determines the size of hyper-cubes around each (high-dimensional) sample point. If smaller than 1, this is computed as a fraction of the sample size as knn=n*knn. For knn larger or equal to 1, knn=knn}

\item{shuffle_neighbors}{Number of nearest-neighbors within Z for the shuffle surrogates which determines the size of hyper-cubes around each (high-dimensional) sample point.}

\item{sig_samples}{Number of samples for shuffle significance test.}

\item{transform}{(default: ‘ranks’) Whether to transform the array beforehand by standardizing or transforming to uniform marginals.}
}
\value{
list of p-value and the conditional mutual information of X,Y given Z
}
\description{
Conditional independence test (CIT) based on the conditional mutual information (CMI) with k-nearest-neighbors (k-NN) estimation
}
\examples{
\dontshow{if (reticulate::py_module_available('tigramite')) withAutoprint(\{ # examplesIf}
n <- 20; p <- 1; q <- 1
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
Y <- matrix(rnorm(n), nrow = n)
Z <- matrix(rnorm(n*q), nrow = n, ncol = q)
cmiknn(X, Y, Z)
\dontshow{\}) # examplesIf}
}
