% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcor_cpt.R
\name{dcor_cpt}
\alias{dcor_cpt}
\title{Conditional permutation-distance correlation conditional independence test}
\usage{
dcor_cpt(X, Y, Z, index = 1, model.formula.YZ, M = 30, nstep = 50)
}
\arguments{
\item{X}{A nxp-matrix of n p-dimensional vectors}

\item{Y}{A nx1-matrix}

\item{Z}{A nxq-matrix of n q-dimensional confounder (optional)}

\item{index}{exponent of Euclidean distance used in distance correlation (default: 1)}

\item{model.formula.YZ}{A formula for the generalized additive model approximating the conditional distribution of Y given Z}

\item{M}{Number of conditional permutations}

\item{nstep}{Parameter inside MCMC to generate conditional permutations}
}
\value{
list of p-value and distance correlation for observed X,Y,Z
}
\description{
Conditional independence test (CIT) based on the combination of the conditional permutation to control the T1E and
distance correlation as test statistic.
}
\examples{
n <- 20; p <- 10; q <- 2
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
Y <- matrix(rnorm(n), nrow = n)
Z <- matrix(rnorm(n*q), nrow = n, ncol = q)
if(requireNamespace("energy", quietly = TRUE)){
  withr::local_package("energy")
  dcor_cpt(X, Y, Z, model.formula.YZ='V1~V2+V3')
}
}
