% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNCIT-package.R
\docType{package}
\name{DNCIT-package}
\alias{DNCIT-package}
\title{DNCIT: Deep Nonparametric Conditional Independence Test for Images}
\description{
Deep Nonparametric Conditional Independence Test (DNCIT) for Images, based on a user-given embedding map of the image and the application of existing nonparametric CITs to the feature representation of the image and a scalar, given vector-valued confounders.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MSimnach/DNCIT}
  \item Report bugs at \url{https://github.com/MSimnach/DNCIT/issues}
}

}
\author{
\strong{Maintainer}: Marco Simnacher \email{marco.simnacher@hu-berlin.de}

}
\keyword{internal}
