% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpc_graph.R
\name{kpc_graph}
\alias{kpc_graph}
\title{Conditional permutation-kernel partial correlation conditional independence test}
\usage{
kpc_graph(X, Y, Z, k, Knn, model.formula.YZ, M = 30, nstep = 50)
}
\arguments{
\item{X}{A nxp-matrix of n p-dimensional vectors}

\item{Y}{A nx1-matrix}

\item{Z}{A nxq-matrix of n q-dimensional confounder (optional)}

\item{k}{A kernel chosen from available kernel in kernlab}

\item{Knn}{Number of nearest neighbors for KPCgraph}

\item{model.formula.YZ}{A formula for the generalized additive model approximating the conditional distribution of Y given Z}

\item{M}{Number of conditional permutations}

\item{nstep}{Parameter inside MCMC to generate conditional permutations}
}
\value{
list of p-value and kernel partial correlation for observed X,Y,Z
}
\description{
Conditional independence test (CIT) based on the combination of the conditional permutation to control the T1E and
Kernel partial correlation as test statistic.
}
\examples{
n <- 20; p <- 10; q <- 2
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
Y <- matrix(rnorm(n), nrow = n)
Z <- matrix(rnorm(n*q), nrow = n, ncol = q)
if(requireNamespace("kernlab", quietly = TRUE)){
  withr::local_package("kernlab")
  kpc_graph(X, Y, Z, k=kernlab::rbfdot(1/(2 * stats::median(stats::dist(Y))^2)),
  Knn = 1, model.formula.YZ='V1~V2+V3')
}
}
