% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_cit.R
\name{pred_cit}
\alias{pred_cit}
\title{Prediction based CIT with mlr3 and random forest
Auto tuner for hyperparameter optimization
Nested resampling for performance estimation}
\usage{
pred_cit(X, Y, Z, term_time = round(exp(4)/3))
}
\arguments{
\item{X}{A nxp-matrix of n p-dimensional vectors}

\item{Y}{A nx1-matrix}

\item{Z}{A nxq-matrix of n q-dimensional confounder (optional)}

\item{term_time}{The termination time for the auto tuner}
}
\value{
list of p-value and test statistic
}
\description{
Prediction based CIT with mlr3 and random forest
Auto tuner for hyperparameter optimization
Nested resampling for performance estimation
}
\examples{
library(mlr3learners)
n <- 20; p <- 10; q <- 2
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
Y <- matrix(rnorm(n), nrow = n)
Z <- matrix(rnorm(n*q), nrow = n, ncol = q)
pred_cit(X, Y, Z, term_time = 1)
}
