% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_fcit.R
\name{r.fcit}
\alias{r.fcit}
\title{Fast conditional independence test}
\usage{
r.fcit(X, Y, Z)
}
\arguments{
\item{X}{A nxp-matrix of n p-dimensional vectors}

\item{Y}{A nx1-matrix}

\item{Z}{A nxq-matrix of n q-dimensional confounder (optional)}
}
\value{
list of p-value and -1 as test statistic, because the CIT does not return a test statistic
}
\description{
Fast conditional independence test
}
\examples{
n <- 20; p <- 10; q <- 2
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
Y <- matrix(rnorm(n), nrow = n)
Z <- matrix(rnorm(n*q), nrow = n, ncol = q)
fcit(X,Y,Z)
}
