% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_ai_clip.R
\name{r_open_ai_clip}
\alias{r_open_ai_clip}
\title{Open AI CLIP}
\usage{
r_open_ai_clip(
  params_open_ai_clip = NULL,
  dir_path = dir_path,
  all_files = all_files
)
}
\arguments{
\item{params_open_ai_clip}{Specification which open ai clip model to use, default is (ViT-B-32, pretrained=laion2bs34bb79k)}

\item{dir_path}{Path to the directory containing the images}

\item{all_files}{List of all files in the directory}
}
\value{
Feature representation of the image obtained by open ai model
}
\description{
Open AI CLIP
}
