% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tucker_decomposition.R
\name{tucker_decomposition}
\alias{tucker_decomposition}
\title{Tucker decomposition}
\usage{
tucker_decomposition(
  params_tuckerD = list(dim_reduced = c(10, 10)),
  img_dir_path = NULL
)
}
\arguments{
\item{params_tuckerD}{Parameters dim_reduced of the reduced array dimension (default c(10,10) for 2D images), and max_iter, tol as described in \code{rTensor::tucker()}}

\item{img_dir_path}{Path to the directory with the images}
}
\value{
X Feature representations of the images of dimension as vectorization of dim_reduced array
}
\description{
Tucker decomposition
}
