% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNCIT.R
\name{DNCIT}
\alias{DNCIT}
\title{Title}
\usage{
DNCIT(X, Y, Z, embedding_map = NULL, cit = "RCOT", params_CIT = list())
}
\arguments{
\item{X}{A nxp-matrix of n p-dimensional (vectorized) images or a feature representation of the images}

\item{Y}{A scalar}

\item{Z}{A q-dimensional vector-valued confounder (optional)}

\item{embedding_map}{A embedding map computing feature representations from the images (optional)}

\item{cit}{The nonparametric CIT applied to the feature representations, Y and Z. Default is "RCOT"}

\item{params_CIT}{A list of parameters for the nonparametric CIT}
}
\value{
p-value and test statistic as vector
}
\description{
Title
}
\examples{
n <- 100; p <- 10; q <- 2
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
Y <- rnorm(n)
Z <- matrix(rnorm(n*q), nrow = n, ncol = q)
res <- DNCIT(X, Y, Z)
}
