% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_ai_clip.R
\name{r_open_ai_clip}
\alias{r_open_ai_clip}
\title{Open AI CLIP}
\usage{
r_open_ai_clip(params_open_ai_clip = NULL, img_dir_path = NULL)
}
\arguments{
\item{params_open_ai_clip}{Specification which open ai clip model to use, default is (ViT-B-32, pretrained=laion2bs34bb79k)}

\item{img_dir_path}{Path to the directory containing the images}
}
\value{
Feature representation of the image obtained by open ai model
}
\description{
Open AI CLIP
}
\examples{
\dontshow{if (reticulate::py_module_available('open_clip') && requireNamespace("fs") && reticulate::py_module_available('PIL.image')) withAutoprint(\{ # examplesIf}
path_emp_result_plot <- fs::path_package("extdata/emp_results_plots/", package = "DNCIT")
r_open_ai_clip(img_dir_path = path_emp_result_plot)
\dontshow{\}) # examplesIf}
}
