% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_ai_clip.R
\name{r_open_ai_clip}
\alias{r_open_ai_clip}
\title{Open AI CLIP}
\usage{
r_open_ai_clip(PIL_img, params_open_ai_clip = NULL)
}
\arguments{
\item{PIL_img}{image loaded with PIL.Image.open}

\item{params_open_ai_clip}{Specification which open ai clip model to use, default is (ViT-B-32, pretrained=laion2bs34bb79k)}
}
\value{
Feature representation of the image obtained by open ai model
}
\description{
Open AI CLIP
}
